+++ semgrep-core/src/matching/Generic_vs_generic.ml
@@ -1688,19 +1688,21 @@ and m_other_attribute_operator = m_other_xxx
  * todo? we could restrict ourselves to only a few forms?
  *)
 (* experimental! *)
-and m_stmts_deep ~less_is_ok (xsa : G.stmt list) (xsb : G.stmt list) tin =
+and m_stmts_deep ~inside ~less_is_ok (xsa : G.stmt list) (xsb : G.stmt list) tin
+    =
   (* shares the cache with m_list__m_stmt *)
   match (tin.cache, xsa, xsb) with
   | Some cache, a :: _, _ :: _ when a.s_use_cache ->
       let tin = { tin with mv = Env.update_min_env tin.mv a } in
       Caching.Cache.match_stmt_list ~access:cache_access ~cache
         ~function_id:CK.Match_deep ~list_kind:CK.Original ~less_is_ok
-        ~compute:(m_stmts_deep_uncached ~less_is_ok)
+        ~compute:(m_stmts_deep_uncached ~inside ~less_is_ok)
         ~pattern:xsa ~target:xsb tin
-  | _ -> m_stmts_deep_uncached ~less_is_ok xsa xsb tin
+  | _ -> m_stmts_deep_uncached ~inside ~less_is_ok xsa xsb tin
 
 (*s: function [[Generic_vs_generic.m_stmts_deep]] *)
-and m_stmts_deep_uncached ~less_is_ok (xsa : G.stmt list) (xsb : G.stmt list) =
+and m_stmts_deep_uncached ~inside ~less_is_ok (xsa : G.stmt list)
+    (xsb : G.stmt list) =
   (* opti: this was the old code:
    *   if !Flag.go_deeper_stmt && (has_ellipsis_stmts xsa)
    *   then
@@ -1736,6 +1738,17 @@ and m_stmts_deep_uncached ~less_is_ok (xsa : G.stmt list) (xsb : G.stmt list) =
   | [], _ :: _ -> if less_is_ok then return () else fail ()
   (* dots: '...', can also match no statement *)
   | [ { s = G.ExprStmt ({ e = G.Ellipsis _i; _ }, _); _ } ], [] -> return ()
+  (* inside:
+   * When a pattern-inside (or pattern-not-inside) ends in ... we do not
+   * need to compute all possible endings, it's enough to return the largest
+   * one. If some other pattern P matches inside the largest match here, then
+   * that is enough to either keep P (pattern-inside) or filter it out
+   * (pattern-not-inside).
+   *)
+  | [ { s = G.ExprStmt ({ e = G.Ellipsis _i; _ }, _); _ } ], xb :: bbs
+    when inside ->
+      env_add_matched_stmt xb >>= fun () ->
+      m_stmts_deep_uncached ~inside ~less_is_ok xsa bbs
   | ( ({ s = G.ExprStmt ({ e = G.Ellipsis _i; _ }, _); _ } :: _ as xsa),
       (_ :: _ as xsb) ) ->
       (* let's first try without going deep *)
@@ -1759,7 +1772,8 @@ and m_stmts_deep_uncached ~less_is_ok (xsa : G.stmt list) (xsb : G.stmt list) =
   (* the general case *)
   | xa :: aas, xb :: bbs ->
       m_stmt xa xb >>= fun () ->
-      env_add_matched_stmt xb >>= fun () -> m_stmts_deep ~less_is_ok aas bbs
+      env_add_matched_stmt xb >>= fun () ->
+      m_stmts_deep ~inside ~less_is_ok aas bbs
   | _ :: _, _ -> fail ()
 
 (*e: function [[Generic_vs_generic.m_stmts_deep]] *)
@@ -1920,7 +1934,8 @@ and m_stmt a b =
       or_list m_stmt a bs
   (* the general case *)
   (* ... will now allow a subset of stmts (less_is_ok = false here) *)
-  | G.Block a1, B.Block b1 -> m_bracket (m_stmts_deep ~less_is_ok:false) a1 b1
+  | G.Block a1, B.Block b1 ->
+      m_bracket (m_stmts_deep ~inside:false ~less_is_ok:false) a1 b1
   (*e: [[Generic_vs_generic.m_stmt()]] deep matching cases *)
   (*s: [[Generic_vs_generic.m_stmt()]] builtin equivalences cases *)
   (* equivalence: vardef ==> assign, and go deep *)
@@ -2824,7 +2839,7 @@ and m_partial a b =
 and m_any a b =
   match (a, b) with
   | G.Str a1, B.Str b1 -> m_string_ellipsis_or_metavar_or_default a1 b1
-  | G.Ss a1, B.Ss b1 -> m_stmts_deep ~less_is_ok:true a1 b1
+  | G.Ss a1, B.Ss b1 -> m_stmts_deep ~inside:false ~less_is_ok:true a1 b1
   | G.E a1, B.E b1 -> m_expr a1 b1
   | G.S a1, B.S b1 -> m_stmt a1 b1
   (*s: [[Generic_vs_generic.m_any]] boilerplate cases *)