+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Dataflow: Disregard type arguments but not the entire instruction
 
 ### Changed
+- Optimize ending `...` in `pattern-inside`s to simply match anything left
 
 ## [0.62.0](https://github.com/returntocorp/semgrep/releases/tag/v0.62.0) - 2021-08-17